requirements.txt
PyQt6==6.7.1
yara-python==4.5.1
